+++ previous-versions/banking/src/sdk/sdk.ts
@@ -3,9 +3,9 @@
  */
 
 import * as utils from "../internal/utils";
+import * as shared from "../sdk/models/shared";
 import { AccountBalances } from "./accountbalances";
 import { Accounts } from "./accounts";
-import * as shared from "./models/shared";
 import { TransactionCategories } from "./transactioncategories";
 import { Transactions } from "./transactions";
 import axios from "axios";
@@ -57,9 +57,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.31.0";
-    genVersion = "2.173.0";
-    userAgent = "speakeasy-sdk/typescript 0.31.0 2.173.0 3.0.0 @codat/banking";
+    sdkVersion = "0.32.0";
+    genVersion = "2.188.1";
+    userAgent = "speakeasy-sdk/typescript 0.32.0 2.188.1 3.0.0 @codat/banking";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);
@@ -106,7 +106,7 @@ export class CodatBanking {
             serverURL = ServerList[serverIdx];
         }
 
-        const defaultClient = props?.defaultClient ?? axios.create({ baseURL: serverURL });
+        const defaultClient = props?.defaultClient ?? axios.create();
         this.sdkConfiguration = new SDKConfiguration({
             defaultClient: defaultClient,
             security: props?.security,