+++ previous-versions/banking/RELEASES.md
@@ -559,3 +559,13 @@ Based on:
 - [typescript v0.31.0] previous-versions/banking
 ### Releases
 - [NPM v0.31.0] https://www.npmjs.com/package/@codat/banking/v/0.31.0 - previous-versions/banking
+
+## 2023-11-13 01:24:41
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Banking.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.32.0] previous-versions/banking
+### Releases
+- [NPM v0.32.0] https://www.npmjs.com/package/@codat/banking/v/0.32.0 - previous-versions/banking